[TASK] Suggest change from WTFPL to MIT
We love sticky kit and our developers requested to use it multiple times now. Unfortunately we are currently not able to allow them to use it as WTFPL license is like dedication to the public domain. In some countries in europe law doesnt allow such dedication. Copyright has to be maintained by a licence always. The validity of WTFPL license is disputed by the Open Source Initiative - so we suggest to change license to MIT which you referenced in bower.json anyway.

Maybe you can weigh the pros and cons about such a change - we would be very happy to use your library. I think our doubts are common for web agencies in europe.